requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
docker==5.0.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
pycparser==2.20
pyOpenSSL==20.0.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.1
